+++ client/app/componenets/CourseCreate.js
@@ -1,18 +1,26 @@
+// react imports
 import React, { useState } from "react";
+import { useDispatch } from "react-redux";
+import { createCourse } from "../store/slices/courseSlices";
 
+// bootstrap imports
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
 import Modal from "react-bootstrap/Modal";
 
-const CourseCreate = ({
-  setShow,
-  show,
-  handleCreateSubmit,
-  setname,
-  setsubject,
-  setgradelevel,
-}) => {
+const CourseCreate = ({ setShow, show }) => {
   const handleClose = () => setShow(false);
+  const dispatch = useDispatch();
+
+  const [name, setName] = useState("");
+  const [subject, setSubject] = useState("");
+  const [gradeLevel, setGradeLevel] = useState("");
+
+  const handleCreateSubmit = (e) => {
+    e.preventDefault();
+    dispatch(createCourse({ name, subject, gradeLevel }));
+    setShow(false);
+  };
 
   return (
     <Modal show={show} onHide={handleClose}>
@@ -27,8 +35,7 @@ const CourseCreate = ({
               type="text"
               placeholder="Class Name"
               autoFocus
-              ref={setname}
-              onChange={(e) => setname(e.target.value)}
+              onChange={(e) => setName(e.target.value)}
             />
           </Form.Group>
           <Form.Group controlId="">
@@ -36,8 +43,7 @@ const CourseCreate = ({
             <Form.Control
               type="text"
               placeholder="Subject Name"
-              ref={setsubject}
-              onChange={(e) => setsubject(e.target.value)}
+              onChange={(e) => setSubject(e.target.value)}
             />
           </Form.Group>
 
@@ -46,8 +52,7 @@ const CourseCreate = ({
             <Form.Control
               type="text"
               placeholder="Grade"
-              ref={setgradelevel}
-              onChange={(e) => setgradelevel(e.target.value)}
+              onChange={(e) => setGradeLevel(e.target.value)}
             />
           </Form.Group>
         </Form>
