+++ client/app/componenets/CourseEdit.js
@@ -1,22 +1,15 @@
+//react stuff
 import React, { useState } from "react";
+import { useDispatch } from "react-redux";
+import { editCourse } from "../store/slices/courseSlices";
 
+//bootstrap stuff
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
 import Modal from "react-bootstrap/Modal";
-import { useDispatch, useSelector } from "react-redux";
-import { editCourse } from "../store/slices/courseSlices";
-import { useParams, Link } from "react-router-dom";
 
-const CourseCreate = ({
-  setShowEdit,
-  showEdit,
-  setname,
-  setsubject,
-  setgradelevel,
-}) => {
-  const handleClose = () => setShowEdit(false);
+const CourseCreate = ({ showEdit, setShowEdit, id }) => {
   const dispatch = useDispatch();
-  const { courseId } = useParams();
 
   const [name, setName] = useState("");
   const [subject, setSubject] = useState("");
@@ -24,13 +17,16 @@ const CourseCreate = ({
 
   const handleEditSubmit = (e) => {
     e.preventDefault();
-    dispatch(editCourse({ courseId, name, subject, gradeLevel }));
+    dispatch(editCourse({ id, name, subject, gradeLevel }));
     setShowEdit(false);
+    console.log(showEdit, setShowEdit, id);
   };
+
+  const handleEditClose = () => setShowEdit(false);
 
   return (
     <>
-      <Modal show={showEdit} onHide={handleClose}>
+      <Modal show={showEdit} onHide={handleEditClose}>
         <Modal.Header closeButton>
           <Modal.Title>Edit Class</Modal.Title>
         </Modal.Header>
@@ -42,7 +38,6 @@ const CourseCreate = ({
                 type="text"
                 placeholder="Class Name"
                 autoFocus
-                ref={setname}
                 onChange={(e) => setName(e.target.value)}
               />
             </Form.Group>
@@ -51,7 +46,6 @@ const CourseCreate = ({
               <Form.Control
                 type="text"
                 placeholder="Subject Name"
-                ref={setsubject}
                 onChange={(e) => setSubject(e.target.value)}
               />
             </Form.Group>
@@ -61,14 +55,13 @@ const CourseCreate = ({
               <Form.Control
                 type="text"
                 placeholder="Grade"
-                ref={setgradelevel}
                 onChange={(e) => setGradeLevel(e.target.value)}
               />
             </Form.Group>
           </Form>
         </Modal.Body>
         <Modal.Footer>
-          <Button variant="secondary" onClick={handleClose}>
+          <Button variant="secondary" onClick={handleEditClose}>
             Close
           </Button>
           <Button variant="primary" type="submit" onClick={handleEditSubmit}>
