+++ client/app/screen/CourseScreen.js
@@ -7,26 +7,15 @@ import Form from "react-bootstrap/Form";
 
 //React related imports
 import { useDispatch, useSelector } from "react-redux";
-import {
-  selectCourses,
-  fetchAllCourses,
-  createCourse,
-  editCourse,
-} from "../store/slices/courseSlices";
-import { useParams, Link } from "react-router-dom";
+import { selectCourses, fetchAllCourses } from "../store/slices/courseSlices";
 import CourseCreate from "../componenets/CourseCreate";
 import CourseEdit from "../componenets/CourseEdit";
-import CourseStudentScreen from "./CourseStudentScreen";
 
 const CousreScreen = () => {
   const [show, setShow] = useState(false);
   const [showEdit, setShowEdit] = useState(false);
-  const [name, setName] = useState("");
-  const [subject, setSubject] = useState("");
-  const [gradeLevel, setGradeLevel] = useState("");
 
   const courses = useSelector(selectCourses);
-  console.log(courses);
   const dispatch = useDispatch();
 
   useEffect(() => {
@@ -35,18 +24,6 @@ const CousreScreen = () => {
 
   const handleShow = () => setShow(true);
   const handleShowEdit = () => setShowEdit(true);
-
-  const handleCreateSubmit = (e) => {
-    e.preventDefault();
-    dispatch(createCourse({ name, subject, gradeLevel }));
-    setShow(false);
-  };
-
-  // const handleEditSubmit = (e) => {
-  //   e.preventDefault();
-  //   dispatch(editCourse({ name, subject, gradeLevel }));
-  //   setShow(false);
-  // };
 
   return (
     <div>
@@ -67,52 +44,50 @@ const CousreScreen = () => {
         </thead>
         <tbody>
           {courses && courses.length
-            ? courses.map((course) => {
-                return (
-                  <tr key={course.id}>
-                    <td>
-                      <Form.Check type="checkbox" />
-                    </td>
-                    <td>{course.name}</td>
-                    <td>{course.subject}</td>
-                    <td>{course.gradeLevel}</td>
-                    <td>
-                      <Dropdown>
-                        <Dropdown.Toggle id="dropdown-basic"></Dropdown.Toggle>
-                        <Dropdown.Menu>
-                          <Dropdown.Item
-                            href={`/courses/${course.id}/students`}
-                          >
-                            Students
-                          </Dropdown.Item>
-                          <Dropdown.Item>Assessments</Dropdown.Item>
-                          <Dropdown.Item>Report</Dropdown.Item>
-                          <Dropdown.Item onClick={handleShowEdit}>
-                            Edit
-                          </Dropdown.Item>
-                          <Dropdown.Item>Archive</Dropdown.Item>
-                        </Dropdown.Menu>
-                      </Dropdown>
-                    </td>
-                  </tr>
-                );
-              })
+            ? courses
+                .filter((remainingCourse) => remainingCourse.id !== courses.id)
+                .map((course) => {
+                  return (
+                    <tr key={course.id}>
+                      <td>
+                        <Form.Check type="checkbox" />
+                      </td>
+                      <td>{course.name}</td>
+                      <td>{course.subject}</td>
+                      <td>{course.gradeLevel}</td>
+                      <td>
+                        <Dropdown>
+                          <Dropdown.Toggle id="dropdown-basic"></Dropdown.Toggle>
+                          <Dropdown.Menu>
+                            <Dropdown.Item
+                              href={`/courses/${course.id}/students`}
+                            >
+                              Students
+                            </Dropdown.Item>
+                            <Dropdown.Item>Assessments</Dropdown.Item>
+                            <Dropdown.Item>Report</Dropdown.Item>
+                            <Dropdown.Item onClick={handleShowEdit}>
+                              Edit
+                            </Dropdown.Item>
+                            <Dropdown.Item>Archive</Dropdown.Item>
+                          </Dropdown.Menu>
+                          <CourseEdit
+                            showEdit={showEdit}
+                            setShowEdit={setShowEdit}
+                            id={course.id}
+                          />
+                        </Dropdown>
+                      </td>
+                    </tr>
+                  );
+                })
             : null}
         </tbody>
       </Table>
 
       <Button variant="danger">Archive</Button>
 
-      <CourseCreate
-        show={show}
-        setShow={setShow}
-        handleCreateSubmit={handleCreateSubmit}
-        setname={setName}
-        setsubject={setSubject}
-        setgradelevel={setGradeLevel}
-      />
-
-      <CourseEdit showEidt={showEdit} setShowEdit={setShowEdit} />
+      <CourseCreate show={show} setShow={setShow} />
     </div>
   );
 };
